A Leaflet/D3 demo for drones.
This simple Leaflet map allows you to randomly place a specified number
of waypoint drones on the map. You can then drag-and-drop a (draggable)
mission drone on to the map, which will display D3 lines of varying
color between the mission and way point drones. This lines are a dynamic
indication of a somewhat arbitrary prioritization of distances between
mission and waypoint drones.